+++ src/environments/environment.vegas.ts
@@ -1,4 +1,6 @@
-export const environment = {
+import { Environment } from "./environment.model";
+
+export const environment: Environment = {
   production: true,
   simulatedTime: undefined,
   gps: { lng: -119.19588439718973, lat: 40.78098917543007 },